Fixed group script editing in admin.php frontend/php.

+++ frontend/php/admin_changes.php
@@ -2,16 +2,44 @@
 $dir = "./library/";
 $library = scandir($dir);
 $count = count($library);
-for ($x = 0; $x < $count; $x++) {
-	if ($library[$x] != "." && $library[$x] != "..") {
-		$post = $_POST[$library[$x]];
-		if ($post == "1") {
-			// Step up one dir from anticipated symlink target.
-			symlink("../library/$library[$x]", "./scripts/$library[$x]");
-		} elseif($post != "1") {
-			unlink("./scripts/$library[$x]");
+$submit = $_POST['submit'];
+if ($submit == 'Submit') {
+	for ($x = 0; $x < $count; $x++) {
+		if ($library[$x] != "." && $library[$x] != "..") {
+			$post = $_POST[$library[$x]];
+			if ($post == "1") {
+				// Step up one dir from anticipated symlink target.
+				symlink("../library/$library[$x]", "./scripts/$library[$x]");
+			} elseif($post != "1") {
+				unlink("./scripts/$library[$x]");
+			}
 		}
 	}
+echo "admin_change";
+} else {
+	$edit_file = $_POST['edit_file'];
+	$edit_script = $_POST['edit_script'];
+	echo $edit_script."<br />";
+	$edit_script = preg_replace("/\r\n/", "\n", $edit_script); // DOS style newlines
+	$edit_script = preg_replace("/\r/", "\n", $edit_script); // Mac newlines for nostalgia
+	$machine = scandir("./machines/$mgroup/");
+	$count_machines = count($machine);
+	if ($submit == "Submit Changes") {
+		file_put_contents("./library/$edit_file", $edit_script);
+	} else {
+		unlink("./library/$edit_file");
+		if (is_link("./scripts/$edit_file")) {
+			unlink("./scripts/$edit_file");
+		}
+		for ($x = 0; $x < $count_machines; $x++) {
+			if ($machine[$x] != "." && $machine[$x] != "..") {
+				if (is_link("./machines/$machine[$x]/$edit_file")) {
+					unlink("./machines/$machine[$x]/$edit_file");
+				}
+			}
+		}
+	}
+echo "edit";
 }
 header('Location: admin.php');
 ?>
